+++ packages/banana-react/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @banana/banana-react
 
+## 1.10.0-beta.0
+
+### Minor Changes
+
+- Beta version for form, input, rating.
+
+### Patch Changes
+
+- Updated dependencies
+  - banana-ui@1.10.0-beta.0
+
+## 1.9.3
+
+### Patch Changes
+
+- Fixed a bug at autoHeight parameter of carousel component.
+- Updated dependencies
+  - banana-ui@1.9.3
+
+## 1.9.2
+
+### Patch Changes
+
+- Added autoHeight parameter to carousel.
+- Updated dependencies
+  - banana-ui@1.9.2
+
 ## 1.9.1
 
 ### Patch Changes